luci-mod-admin-full: fix connection page
I notice that in the connection tab the lookup array keep checking all the entities everytime and adds to the dns_cache list only the resolved ones. This looks wrong as it's very unluckily that an addr thats not resolved in this instance gets resolved in the next one. This change how the dns_cache list works. We put in it every tried addr even if not resolved. Also this fix some error if for some reason the page doesn't receive the right data. This also fix error with underfined port with icmp traffic type and adds info about the type of connection established. Also sync the table style with the new implementation.

Signed-off-by: Ansuel Smith <ansuelsmth@gmail.com>